reduce stateless-core
