+++ crates/stateless-common/src/lib.rs
@@ -1,7 +1,9 @@
 pub mod db;
 pub mod logging;
+pub mod metrics;
+pub use metrics::{RpcClientConfig, RpcMethod, RpcMetrics};
 pub mod rpc_client;
-pub use rpc_client::RpcClient;
+pub use rpc_client::{RpcClient, SetValidatedBlocksResponse, WitnessRequestKeys};
 
 /// Default port for Prometheus metrics HTTP endpoint.
 pub const DEFAULT_METRICS_PORT: u16 = 9090;
