+++ crates/stateless-core/Cargo.toml
@@ -20,7 +20,6 @@ alloy-primitives.workspace = true
 alloy-rlp.workspace = true
 alloy-rpc-types-eth.workspace = true
 alloy-serde.workspace = true
-alloy-trie.workspace = true
 
 # mega
 mega-evm.workspace = true
